Added autoflush to output filehandles.
Added autoflush to filehandles in order to instantly save to the census every wiki info we retrieve.
This prevents to restart over again when find an error as well as it makes easier to backtrace where the problem was. It also allows to see 'realtime' output of what the script is producing.
IO Waits are not an issue since we want some delay between http request in order to not saturate de server.